pyproject.toml
[project]
name = "cocos-mcp"
version = "1.0.0"
description = "Headless Cocos Creator 3.8 MCP server — 71 tools for AI-autonomous game development"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [{name = "heitugongzuoshi", email = "2282059276@qq.com"}]
keywords = ["cocos", "cocos-creator", "mcp", "game-dev", "headless", "ai"]
dependencies = [
    "mcp[cli]>=1.2.0",
    "Pillow>=10.0.0",
]

[project.urls]
Repository = "https://gitee.com/csbcsb/cocos-mcp"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
